requirements.txt
Flask==2.2.2
requests==2.31.0
python-dotenv==1.0.0
flask-cors==3.0.10
boto3==1.28.0
